requirements.txt
deepgram-sdk
flask